SOCINT-266 Use /survey endpoint for available fulfilments Completed split out of validators
