+++ tests/unit/helpers.py
@@ -350,7 +350,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_enter_address(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.service_calls.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
 
             response = await self.client.request('POST',
@@ -368,7 +368,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_enter_address_finder(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, mock.patch(
-                'app.comms.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
+                'app.service_calls.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             if region == 'W':
                 mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_w
             else:
@@ -389,7 +389,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_enter_address_input_returns_no_results(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.service_calls.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_no_results
 
             response = await self.client.request('POST',
@@ -450,7 +450,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_select_address_no_selection_made(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.service_calls.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
 
             response = await self.client.request('POST',
@@ -467,8 +467,8 @@ class TestHelpers(RHTestCase):
 
     async def check_post_select_address(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
-                'app.comms.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
+                mock.patch('app.service_calls.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
+                'app.service_calls.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
             if region == 'W':
                 mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_w
@@ -558,8 +558,8 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_invalid(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
-                'app.comms.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
+                mock.patch('app.service_calls.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
+                'app.service_calls.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
             mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_e
 
@@ -574,8 +574,8 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_no_selection(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
-                'app.comms.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
+                mock.patch('app.service_calls.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
+                'app.service_calls.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
             mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_e
 
@@ -590,7 +590,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_no(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.service_calls.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
 
             response = await self.client.request('POST',
@@ -609,7 +609,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_yes_code(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Surveys.get_survey_details') as mocked_get_survey_details:
+                mock.patch('app.service_calls.rhsvc.Surveys.get_survey_details') as mocked_get_survey_details:
             mocked_get_survey_details.return_value = self.rhsvc_get_surveys
 
             response = await self.client.request('POST',
@@ -666,7 +666,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_select_how_to_receive_input_invalid(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Surveys.get_survey_details') as mocked_get_survey_details:
+                mock.patch('app.service_calls.rhsvc.Surveys.get_survey_details') as mocked_get_survey_details:
             mocked_get_survey_details.return_value = self.rhsvc_get_surveys
 
             response = await self.client.request('POST',
@@ -681,7 +681,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_select_how_to_receive_input_no_selection(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Surveys.get_survey_details') as mocked_get_survey_details:
+                mock.patch('app.service_calls.rhsvc.Surveys.get_survey_details') as mocked_get_survey_details:
             mocked_get_survey_details.return_value = self.rhsvc_get_surveys
 
             response = await self.client.request('POST',
@@ -772,9 +772,9 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_text(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, mock.patch(
-                'app.comms.rhsvc.Fulfilments.get_fulfilment'
+                'app.service_calls.rhsvc.Fulfilments.get_fulfilment'
         ) as mocked_get_fulfilment, mock.patch(
-            'app.comms.rhsvc.Fulfilments.request_fulfilment_sms'
+            'app.service_calls.rhsvc.Fulfilments.request_fulfilment_sms'
         ) as mocked_request_fulfilment_sms:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_multi_sms
@@ -839,7 +839,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_text_error_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_sms
             mocked_aioresponses.post(self.rhsvc_cases_url +
@@ -856,7 +856,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_text_error_429_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_sms
@@ -907,9 +907,9 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_email(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, mock.patch(
-                'app.comms.rhsvc.Fulfilments.get_fulfilment'
+                'app.service_calls.rhsvc.Fulfilments.get_fulfilment'
         ) as mocked_get_fulfilment, mock.patch(
-            'app.comms.rhsvc.Fulfilments.request_fulfilment_email'
+            'app.service_calls.rhsvc.Fulfilments.request_fulfilment_email'
         ) as mocked_request_fulfilment_email:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_multi_email
@@ -974,7 +974,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_email_error_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_email
             mocked_aioresponses.post(self.rhsvc_cases_url +
@@ -991,7 +991,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_email_error_429_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_email
@@ -1203,8 +1203,8 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_post_input_yes(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
-                mock.patch('app.comms.rhsvc.Fulfilments.request_fulfilment_post') as mocked_request_fulfilment_post:
+                mock.patch('app.service_calls.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.Fulfilments.request_fulfilment_post') as mocked_request_fulfilment_post:
 
             if display_region == 'cy':
                 mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_multi_post
@@ -1285,7 +1285,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_post_error_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_post
             mocked_aioresponses.post(self.rhsvc_cases_url +
@@ -1302,7 +1302,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_post_error_429_from_request_fulfilment_uac(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.comms.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.Fulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_post
